+++ pennylane_lightning/lightning_gpu/lightning_gpu.py
@@ -88,6 +88,29 @@ _operations = frozenset(
         "PauliZ",
         "MultiRZ",
         "GlobalPhase",
+        "C(PauliX)",
+        "C(PauliY)",
+        "C(PauliZ)",
+        "C(Hadamard)",
+        "C(S)",
+        "C(T)",
+        "C(PhaseShift)",
+        "C(RX)",
+        "C(RY)",
+        "C(RZ)",
+        "C(Rot)",
+        "C(SWAP)",
+        "C(IsingXX)",
+        "C(IsingXY)",
+        "C(IsingYY)",
+        "C(IsingZZ)",
+        "C(SingleExcitation)",
+        "C(SingleExcitationMinus)",
+        "C(SingleExcitationPlus)",
+        "C(DoubleExcitation)",
+        "C(DoubleExcitationMinus)",
+        "C(DoubleExcitationPlus)",
+        "C(MultiRZ)",
         "C(GlobalPhase)",
         "Hadamard",
         "S",
@@ -138,6 +161,30 @@ _operations = frozenset(
 )
 # End the set of supported operations.
 
+# TODO: _unsupported_adjoint_ops is a temporary solution to avoid adjoint differentiation for N-controlled gates.
+# This will be removed once the  N-controlled genenerators are implemented.
+_unsupported_adjoint_ops = frozenset(
+    {
+        "C(PhaseShift)",
+        "C(RX)",
+        "C(RY)",
+        "C(RZ)",
+        "C(Rot)",
+        "C(IsingXX)",
+        "C(IsingXY)",
+        "C(IsingYY)",
+        "C(IsingZZ)",
+        "C(SingleExcitation)",
+        "C(SingleExcitationMinus)",
+        "C(SingleExcitationPlus)",
+        "C(DoubleExcitation)",
+        "C(DoubleExcitationMinus)",
+        "C(DoubleExcitationPlus)",
+        "C(MultiRZ)",
+        "C(GlobalPhase)",
+    }
+)
+
 # The set of supported observables.
 _observables = frozenset(
     {
@@ -161,15 +208,6 @@ _observables = frozenset(
 
 def stopping_condition(op: Operator) -> bool:
     """A function that determines whether or not an operation is supported by ``lightning.gpu``."""
-    # To avoid building matrices beyond the given thresholds.
-    # This should reduce runtime overheads for larger systems.
-    if isinstance(op, qml.QFT):
-        return len(op.wires) < 10
-    if isinstance(op, qml.GroverOperator):
-        return len(op.wires) < 13
-    if isinstance(op, qml.PauliRot):
-        return False
-
     return op.name in _operations
 
 
@@ -225,7 +263,15 @@ def _supports_adjoint(circuit):
 
 def _adjoint_ops(op: qml.operation.Operator) -> bool:
     """Specify whether or not an Operator is supported by adjoint differentiation."""
-    return not isinstance(op, qml.PauliRot) and adjoint_ops(op)
+    # FIXME: This is a temporary solution to avoid adjoint differentiation for N-controlled gates.
+    if op.name in _unsupported_adjoint_ops:
+        # "C(SingleExcitation)" is not supported by the lightning.gpu after decomposition.
+        if op.name == "C(SingleExcitation)":
+            raise qml.DeviceError(
+                "C(SingleExcitation) is not supported by adjoint differentiation."
+            )
+        return False
+    return adjoint_ops(op) and not isinstance(op, qml.PauliRot)
 
 
 def _add_adjoint_transforms(program: TransformProgram) -> None: